feat: add new list button
This one feature was a little tough to do. To create a list we need to
first display dialog and then call save method. Save method must be called
with cur.tab_hash which is unfortunately only provided on feed page, so
we must redirect user to feed page.

Here nav.go comes in handy which allows AJAX pages navigation with callback
onDone, so we can pop up the creation dialog, using feed.editList method.
Unfortunately again, we cannot normally add and select pages to list.

If we debug dialog code, we can find that dialog box is calling some OList
object in `window`, turns out it is a controller for the list, which
is called with arguments such as list, so we can fake it and change list
to add element and automatically select needed item.

To make it user-friendly, we are highlighting the item and scrolling to it.