refactor: deprecate away VK_API.isUsingRuLocale
VK reloads page if language is changed and there is no practical sense
to go over and over array again every time we need to check used locale.
So it has sense to replace it with something permanent - like a symbol.

Unfortunately we have no ways we can set it on page load, because if
we wait for page to load first, then menuItem / actionButton mount will
fail because it's called earlier than page load event is fired.

With a little of improvisation we can use something like memo function --

We still internally have this function but now it's replaced with symbol
RU_LOCALE_USED, which on first use will call isUsingRuLocale and then
will rewrite itself with permanent value, returning it to caller. Hacky.