refactor: replace SYMBOLS collection with SYM__* constants
The use of SYMBOLS collection negates the performance of symbols lookups
and makes it useless to use symbols there in the first place. I agree
that usage of symbols is questionable at the first place in some cases,
but I want to at least try optimizing them before removing completely.

Plus, other that that, symbols give us the idea that a property is not
for general use (in most cases). And if we look at them like this there
a lot more work to be done while refactoring ...

But that's all we got for now in this particular commit.

Thanks for coming to my TED talk.