feat: own styles for info block
On some pages modules.css is not being loaded, but it contains styles
needed for us. We could possibly load it using stManager.add, but there
are two downsides I can think of:

- First is bandwidth - we will be required to download a big CSS file
  just for this simple dialog.

- The second is performance - I believe stManager.add is not very
  optimized, plus VKOpt wraps it in order to fire events about resource
  loading. So, overhead.

- And last one is additional callback, which we will be required to
  promisify, what is not very nice practice.

And by copying we have benefits like being able to issue fixes, but sadly
also other styles won't be changing it's style because we must use
different selectors in order to not break things. But that's all.

So I went with copying CSS in this commit.